+++ crates/mimic_base/src/lib.rs
@@ -13,6 +13,6 @@ pub mod validator;
 pub(crate) mod prelude {
     pub use crate::{types, validator};
     pub use mimic::prelude::*;
-    pub use mimic::types::prim::*;
+    pub use mimic::types::*;
     pub use mimic_design::*;
 }